Update README.md
This is a data mining project about world tourism analysis.

The international travel & tourism sector is growing at a fast pace. The World Travel and Tourism Council (WTTC) report found that the travel & tourism industry generated 10.4 percent of all global economic growth. Tourism is not only a kind of leisure, it is also essential in a country's economy. In this analysis, I investigate the dynamic trend of evolvement in tourism sector and find the top tourist-sending countries and the countries that are most attractive to international tourists.

Language:
python

packages:
pandas
numpy
plotly
plotly_express

Files:
Excel
ipynb
html

Instruction:
Put the excel and python code in the same folder
Need plotly account to run the code
Recommend use google chrome